# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yalla Wassel delivery management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Phone already registered
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /customer/orders:
    post:
      operationId: createCustomerOrder
      tags: [orders]
      summary: Create an order as a customer (auth required, customer role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
        "400":
          description: Validation error
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get current customer's orders
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithDetails"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /ai/suggest-driver:
    post:
      operationId: suggestDriver
      tags: [ai]
      summary: AI-powered driver suggestion for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverSuggestionRequest"
      responses:
        "200":
          description: Suggested driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverSuggestionResponse"
  /ai/detect-zone:
    post:
      operationId: detectZone
      tags: [ai]
      summary: AI zone detection from address text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneDetectionRequest"
      responses:
        "200":
          description: Detected zone suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetectionResponse"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [fullName, phone, password, role]
      properties:
        fullName:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [manager, driver, customer]
        zoneId:
          type: ["integer", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Zone schemas ───────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, phone, role, status, isActive, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [manager, driver, customer]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, off]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    OrderWithDetails:
      type: object
      required: [id, orderId, fromBusiness, fromAddress, fromZoneId, toCustomerName, toPhone, toAddress, toZoneId, priority, status, createdAt, trackingToken]
      properties:
        id:
          type: integer
        orderId:
          type: string
        fromBusiness:
          type: string
        fromAddress:
          type: string
        fromZoneId:
          type: integer
        fromZoneName:
          type: ["string", "null"]
        toCustomerName:
          type: string
        toPhone:
          type: string
        toAddress:
          type: string
        toZoneId:
          type: integer
        toZoneName:
          type: ["string", "null"]
        priority:
          type: string
          enum: [urgent, normal]
        status:
          type: string
          enum: [pending, assigned, picked, delivered, cancelled]
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        driverPhone:
          type: ["string", "null"]
        driverAvgRating:
          type: ["number", "null"]
        customerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        assignedAt:
          type: ["string", "null"]
        pickedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        eta:
          type: ["string", "null"]
        trackingToken:
          type: string
...
    CustomerOrderInput:
      type: object
      required: [fromBusiness, fromAddress, fromZoneId, toAddress, toZoneId, category]
      properties:
        category:
          type: string
          enum: [food, flowers, pharmacy, grocery, electronics, clothing, other]
        fromBusiness:
          type: string
        fromAddress:
          type: string
        fromZoneId:
          type: integer
        toAddress:
          type: string
        toZoneId:
          type: integer
        notes:
          type: string
...
    Notification:
      type: object
      required: [id, userId, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [order, system, rating]
        isRead:
          type: boolean
        createdAt:
          type: string

    # ── Dashboard schemas ─────────────────────────────────────────────────────
...
    DriverSuggestionRequest:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: integer
    DriverSuggestionResponse:
      type: object
      required: [driverId, driverName, reason]
      properties:
        driverId:
          type: integer
        driverName:
          type: string
        reason:
          type: string
    ZoneDetectionRequest:
      type: object
      required: [address]
      properties:
        address:
          type: string
    ZoneDetectionResponse:
      type: object
      required: [zoneId, zoneName, confidence]
      properties:
        zoneId:
          type: integer
        zoneName:
          type: string
        confidence:
          type: string