# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yalla Wassel delivery management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  zoneName?: string | null;
  status: UserStatus;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}